seo_agent-main/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
aiohttp==3.9.1
pydantic==2.5.0
groq==0.4.2
anthropic==0.7.0
openai==1.3.9
